refactor(#147): extract wallet nodes from apidata.js, share REQUEST_OPTIONS_API
apidata.js 1,055 -> 741 lines (1,702 at the start of #147).

## What moved

client/src/api/walletNodes.js takes the whole "node health" + "nodes overview"
region: getWalletNodes, getEnterpriseNodes, transformRawNode, fillPartialNode,
fill_health, wallet_health_full, normalize_raw_node_tier, calc_mtn_window,
validateAddress, getDemoWallet, and DISPLAY_DATE_FORMAT.

The cut was chosen by measuring coupling, not by where the banner comments fell.
Before moving anything: none of those ten functions is called from anywhere else
in apidata.js, and the four endpoint constants they use
(API_FLUX_NODE_URL, API_NODE_INFO_ENDPOINT, API_FLUX_APPLIST_ENDPOINT,
API_FLUX_UPTIME_ENDPOINT) are referenced nowhere outside the region, so they
moved too.

Verified afterwards that the moved block is byte-identical to what was on main.
This is a MOVE, not a rewrite. apidata.js re-exports everything, so no call site
changes in this pass.

## The shared constant, found by the compiler

REQUEST_OPTIONS_API was declared beside the node code but used by
fetch_gpu_prices as well, which stayed behind. Extracting the region broke the
build:

    Line 704:54:  'REQUEST_OPTIONS_API' is not defined  no-undef

It now lives in api/endpoints.js, the module that already exists for things
shared by more than one API module, and both callers import it. A one-constant
module would have been over-splitting.

Worth noting this is exactly the failure #147 is about, in miniature: a value
used from two places, declared next to one of them, where moving either half
breaks the other. And it is the second time this session that a green test run
did not catch a broken build -- the tests pass either way, because none of them
exercises fetch_gpu_prices.

## Also cleaned up

Four imports orphaned by this move (dayjs, format_minutes,
EXPLORER_FLUX_NODES_PATH, FLUXNODE_INFO_API_MODE/URL) and two that were already
dead before it (fluxos_version_string, topInGroup -- left behind by the earlier
#147 passes).

Issue #204: FLUX_PER_DAY's comment said "1 flux every 2 minutes", which would be
720. The value is (24 * 60) * 2 = 2,880 -- blocks per day at the 30-second block
target, which is what the CLC_NETWORK_* lines multiply a per-block reward by.
The arithmetic was always right; only the comment was wrong. Corrected rather
than carried forward, since this pass moves the constant's neighbours out from
under it. No behaviour change, and nothing to do with the whitepaper items the
rest of that issue range covers.

## Verification

553 tests pass, production build clean, moved block confirmed byte-identical.
